separate common DB types from insert logic
