+++ stats/stats/src/charts/counters/average_block_time.rs
@@ -1,6 +1,6 @@
 use crate::{
     charts::db_interaction::{
-        insert::{DateValue, DateValueDouble},
+        types::{DateValue, DateValueDouble},
         updater::ChartFullUpdater,
     },
     UpdateError,
