+++ stats/stats/src/charts/counters/last_new_verified_contracts.rs
@@ -2,7 +2,7 @@ use crate::{
     charts::{
         create_chart,
         db_interaction::{
-            insert::DateValue,
+            types::DateValue,
             updater::{last_point, ChartDependentUpdater},
         },
     },
