+++ stats/stats/src/charts/counters/total_native_coin_transfers.rs
@@ -2,7 +2,7 @@ use crate::{
     charts::{
         create_chart,
         db_interaction::{
-            insert::DateValue,
+            types::DateValue,
             updater::{parse_and_sum, ChartDependentUpdater},
         },
     },
