+++ stats/stats/src/charts/db_interaction/mod.rs
@@ -1,4 +1,5 @@
 //! Abstracted interaction with DB
 
 pub mod insert;
+pub mod types;
 pub mod updater;
