+++ stats/stats/src/charts/lines/accounts_growth.rs
@@ -3,7 +3,7 @@ use crate::{
     charts::{
         cache::Cache,
         db_interaction::{
-            insert::{DateValue, DateValueInt},
+            types::{DateValue, DateValueInt},
             updater::ChartFullUpdater,
         },
     },
