+++ stats/stats/src/charts/lines/average_block_rewards.rs
@@ -1,6 +1,6 @@
 use crate::{
     charts::db_interaction::{
-        insert::{DateValue, DateValueDouble},
+        types::{DateValue, DateValueDouble},
         updater::ChartPartialUpdater,
     },
     UpdateError,
