+++ stats/stats/src/lib.rs
@@ -10,7 +10,6 @@ pub use entity;
 pub use migration;
 
 pub use charts::{
-    cache, counters, db_interaction::insert::DateValue, lines, Chart, MissingDatePolicy,
-    UpdateError,
+    cache, counters, db_interaction::types::DateValue, lines, Chart, MissingDatePolicy, UpdateError,
 };
 pub use read::{get_chart_data, get_counters, ReadError};
