+++ stats/stats/src/read.rs
@@ -1,5 +1,5 @@
 use crate::{
-    charts::db_interaction::insert::DateValue, missing_date::get_and_fill_chart, MissingDatePolicy,
+    charts::db_interaction::types::DateValue, missing_date::get_and_fill_chart, MissingDatePolicy,
 };
 use chrono::NaiveDate;
 use entity::{chart_data, charts};
